Arguments support for all session types
