+++ extensions/ProxySessionBackend/src/CygwinProxySessionBackend.ts
@@ -10,6 +10,7 @@ import * as _ from 'lodash';
 import { app } from 'electron';
 
 import {BulkFileHandle, BulkFileState, CommandEntry, ExtensionContext, Logger, Pty, Terminal, SessionConfiguration, Backend, SessionBackend, EnvironmentMap} from 'extraterm-extension-api';
+import { parseArgs } from 'extraterm-args';
 
 import * as SourceDir from './SourceDir';
 import { ProxyPtyConnector, PtyOptions } from './ProxyPty';
@@ -93,7 +94,7 @@ export class CygwinProxySessionBackend implements SessionBackend {
     const {homeDir, defaultShell} = this._getDefaultCygwinConfig(sessionConfig.cygwinPath);
     const shell = sessionConfig.useDefaultShell ? defaultShell : sessionConfig.shell;
 
-    const args = ["-l"];
+    const args = ["-l"].concat(parseArgs(sessionConfig.args));
     
     const ptyEnv = _.cloneDeep(process.env);
     ptyEnv["TERM"] = "xterm-256color";
