+++ extensions/UnixSessionEditor/src/UnixSessionEditorUi.ts
@@ -36,6 +36,12 @@ import Vue from 'vue';
       </datalist>
     </div>
   </div>
+  <div class="form-group">
+    <label for="name" class="col-sm-4 control-label">Arguments:</label>
+    <div class="input-group col-sm-8">
+      <input type="text" class="form-control" name="args" v-model="args">
+    </div>
+  </div>
 </div>
 `
 })
@@ -45,4 +51,5 @@ export class UnixSessionEditorUi extends Vue {
   useDefaultShell: number = 1;
   shellErrorMsg = "";
   etcShells: string[] = [];
+  args: string;
 }
