+++ packages/extraterm-args/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "extraterm-args",
+  "version": "1.0.0",
+  "description": "parse arguments",
+  "author": "Nick Shanny",
+  "license": "MIT",
+  "main": "main.js",
+  "types": "main.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/sedwards2009/extraterm.git"
+  },
+  "devDependencies": {
+    "typescript": "2.8.3"
+  },
+  "scripts": {
+    "run": "false",
+    "build-all": "yarn run build",
+    "build": "tsc"
+  }
+}
