+++ packages/extraterm-args/tsconfig.json
@@ -1,0 +1,28 @@
+{
+    "compilerOptions": {
+        "target": "es6",
+        "module": "commonjs",
+        "declaration": true,
+        "noImplicitAny": false,
+        "noImplicitThis": true,
+        "noImplicitReturns": true,
+        "removeComments": false,
+        "sourceMap": true,
+        "noLib": false,
+        "outDir": ".",
+        "rootDir": "src",
+        "jsx": "react",
+        "experimentalDecorators": true,
+        "baseUrl": ".",
+        "emitDecoratorMetadata": true
+      
+    },
+    "filesGlob": [
+        "./src/*.ts",
+        "./src/test/*.ts"
+    ],
+    "exclude": [
+        "./*.ts",
+        "./test/*.ts"
+    ]
+}
