+++ packages/extraterm-extension-api/main.ts
@@ -418,6 +418,7 @@ export interface SessionConfiguration {
   uuid: string;
   name: string;             // Human readable name for the profile.
   type?: string;            // type - "cygwin", "babun" or "native" ("" means "native")
+  args?: string;            // command line arguments to be passed to command
 }
 
 /**
